Add target parameter to logging functions
In order to turn on logging for a pane when not being inside this pane, its id
has to be explicitly specified in tmux commands.

When the target pane/window/session of a tmux command is not
explicitly given via the -t parameter, the last target is used. E.g.,
attaching to a session sets the current pane/window/session as target.